Update decodeOpaqueId.js
This is required in order to be consistent with: https://github.com/reactioncommerce/api-utils/blob/f7690705c49c4a0d94799ed3deaf6be9ff909f76/lib/encodeOpaqueId.js#L20

Since the entire encoding/decoding of IDs was built for a reason that is no longer valid, we've found it very practical to simply set the environment variable REACTION_SHOULD_ENCODE_IDS to false. This makes development considerably easier, especially for quick API tests with graphQL playground.
But in order for that to work, `decodeOpaqueId()` should also check for the env variable. If we don't do that, `decodeOpaqueId()` can return wrong `namespace` and `id` values.

(Note that `decodeOpaqueId()` did NOT always fail prior to our change. For example, decoding ID `xifjgK9ckbptZ2ew2 ` will yield a null namespace and return the ID as is, whereas decoding ID `dTp6az9ghgoLSGGTo` will yield a namespace equal to `u` and a wrong ID)

Signed-off-by: hrhosni <hae23g@gmail.com>